+++ src/plugin.ts
@@ -1,19 +1,37 @@
-import { logDebug, loadScript, hasScript } from "./utils";
 import pluginConfig from "./config";
+import { hasScript, loadScript, logDebug } from "./utils";
 
 const inBrowser: boolean = typeof window !== "undefined";
+
+export interface VueGtmTrackEventParams {
+  [key: string]: any;
+  event?: any;
+  category?: any;
+  action?: any;
+  label?: any;
+  value?: any;
+  noninteraction?: boolean;
+}
 
 /**
  * Plugin main class
  */
-export default class AnalyticsPlugin {
+export default class VueGtmPlugin {
   constructor(public readonly id: string | string[]) {}
 
-  enabled() {
-    return pluginConfig.enabled;
+  /**
+   * Check if plugin is enabled
+   */
+  enabled(): boolean {
+    return pluginConfig.enabled ?? true;
   }
 
-  enable(val: boolean) {
+  /**
+   * Enable or disable plugin
+   *
+   * @param val state
+   */
+  enable(val: boolean): void {
     pluginConfig.enabled = val;
 
     if (inBrowser && !!val && !hasScript() && pluginConfig.loadScript) {
@@ -33,27 +51,36 @@ export default class AnalyticsPlugin {
     }
   }
 
-  debugEnabled() {
-    return pluginConfig.debug;
+  /**
+   * Check if plugin is in debug mode
+   */
+  debugEnabled(): boolean {
+    return pluginConfig.debug ?? false;
   }
 
-  debug(val: boolean) {
+  /**
+   * Enable or disable debug mode
+   *
+   * @param val state
+   */
+  debug(val: boolean): void {
     pluginConfig.debug = val;
   }
 
-  dataLayer() {
+  dataLayer(): Array<Record<string, any>> | false {
     if (inBrowser && pluginConfig.enabled) {
       return (window.dataLayer = window.dataLayer || []);
     }
     return false;
   }
 
-  trackView(screenName: string, path: string) {
+  trackView(screenName: string, path: string, additionalEventData: Record<string, any> = {}): void {
     logDebug("Dispatching TrackView", { screenName, path });
 
     if (inBrowser && pluginConfig.enabled) {
       let dataLayer = (window.dataLayer = window.dataLayer || []);
       dataLayer.push({
+        ...additionalEventData,
         event: "content-view",
         "content-name": path,
         "content-view-name": screenName,
@@ -69,7 +96,7 @@ export default class AnalyticsPlugin {
     value = null,
     noninteraction = false,
     ...rest
-  } = {}) {
+  }: VueGtmTrackEventParams = {}): void {
     logDebug("Dispatching event", {
       event,
       category,