+++ src/utils.ts
@@ -1,6 +1,5 @@
 import "url-search-params-polyfill";
-import pluginConfig from "./config";
-import { VueGtmUseOptions } from "./types";
+import pluginConfig, { VueGtmUseOptions } from "./config";
 
 /**
  * Console log depending on config debug mode
@@ -19,10 +18,7 @@ export function logDebug(message: string, args: Record<string, any>): void {
  * @param id GTM ID
  * @param params query params object
  */
-export function loadScript(
-  id: string,
-  config: Pick<VueGtmUseOptions, "defer" | "queryParams"> = {}
-): void {
+export function loadScript(id: string, config: Pick<VueGtmUseOptions, "defer" | "queryParams"> = {}): void {
   const win = window,
     doc = document,
     script = doc.createElement("script"),